+++ CHANGELOG.md
@@ -7,12 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.0.0] - 2021-01-04
 
+### Fixed
+
+- Fixed issue with FTP using non jail
+
 ### Changed
 
 - Changed minimum PHP version to 7.3
 - Changed minimum PHPUnit to 9.2
 - Changed defer to 2.0
 - Changed configurable to 2.0
+- Changed FTPEngine to throw StorageException
+
+### Added
+
+- Added NotFoundException for MemoryEngine
 
 ## [2.0.2]  - 2020-10-24